+++ R/merge_coincident.R
@@ -341,9 +341,13 @@ cycle_signed_area <- function(cycle, pool) {
 #'
 #' @examples
 #' x <- wk::as_wkb(c(
-#'   "MULTIPOLYGON (((0 0, 0 1, 0.75 1, 1 0.8, 0.5 0.7, 0.8 0.6, 0.69 0, 0 0), (0.2 0.2, 0.5 0.2, 0.5 0.4, 0.3 0.6, 0.2 0.4, 0.2 0.2)))",
+#'   paste0(
+#'     "MULTIPOLYGON (((0 0, 0 1, 0.75 1, 1 0.8, 0.5 0.7, ",
+#'     "0.8 0.6, 0.69 0, 0 0), (0.2 0.2, 0.5 0.2, ",
+#'     "0.5 0.4, 0.3 0.6, 0.2 0.4, 0.2 0.2)))"),
 #'   "MULTIPOLYGON (((0.69 0, 0.8 0.6, 1.1 0.63, 1.23 0.3, 0.69 0)))"
 #' ))
+#'
 #' pool <- establish_topology(x)
 #' merged <- merge_coincident(pool)
 #' classify_cycles(merged)
@@ -405,9 +409,13 @@ reverse_cycle <- function(cycle) {
 #'
 #' @examples
 #' polygons_with_holes <- wk::as_wkb(c(
-#'   "MULTIPOLYGON (((0 0, 0 1, 0.75 1, 1 0.8, 0.5 0.7, 0.8 0.6, 0.69 0, 0 0), (0.2 0.2, 0.5 0.2, 0.5 0.4, 0.3 0.6, 0.2 0.4, 0.2 0.2)))",
+#'   paste0(
+#'     "MULTIPOLYGON (((0 0, 0 1, 0.75 1, 1 0.8, 0.5 0.7, ",
+#'     "0.8 0.6, 0.69 0, 0 0), (0.2 0.2, 0.5 0.2, ",
+#'     "0.5 0.4, 0.3 0.6, 0.2 0.4, 0.2 0.2)))"),
 #'   "MULTIPOLYGON (((0.69 0, 0.8 0.6, 1.1 0.63, 1.23 0.3, 0.69 0)))"
 #' ))
+#'
 #' x <- establish_topology(polygons_with_holes)
 #' merged <- merge_coincident(x)
 #' hole_points(merged)